refactor(web): extract TaskPageInner to sibling file; expand custom-layout tests
- task-page-content.tsx was 623 lines, over the 600-line ESLint limit.
  Extract TaskPageInner + TaskPageInnerProps + the local helpers
  (RemoteExecutorStatus, toNullable, resolveRemoteExecutor,
  resolveCurrentStepId, buildTaskTopBarProps, buildTaskLayoutProps,
  maybeBuildDebugEntries) into components/task/task-page-inner.tsx
  (290 lines). task-page-content.tsx drops to 348 lines.
- Export useWorkflowStepsMapped, useSessionPanelState, and
  useMergedAgentState from task-page-content.tsx so the new file
  can reference their ReturnType via type-only imports without
  pulling in their runtime implementations.
- Add three regression tests for applyCustomLayout (null env,
  maximized, env-switch between scheduling and rAF) to mirror
  the applyBuiltInPreset coverage. Hoist customLayout fixture
  and ApplyCustomLayoutArg type alias so all 4 tests share them.
